requirements.txt
asgiref==3.3.0
certifi==2020.6.20
Django==3.1.3
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
wincertstore==0.2
